Catch possible I/O error initializing trust DB
This issue was identified in jupyter/jupyter#174, where takluyver suggested something like this solution.  If the `init_db` fails with an `OperationalError`, this assumes that the problem is an I/O error, and tries to initialize the DB again, using the `:memory:` location so that `sqlite` simply creates an in-memory database.