Try recreating signatures db file for either type of error
Takluyver suggested that these errors may not be sufficiently segregating to clearly partition into (1) cases where we should try to move the db file and recreate it, and (2) cases where we should use memory.  So this simply catches both types of error, and tries (1) first, then again catches both types of error if that fails and tries (2).