Allow notebook format 4.0 to have arbitrary JSON in mimebundle keys ending with +json.
See https://github.com/jupyter/nbformat/pull/167 for more discussion. Essentially, there are a lot of notebooks in the wild that have arbitrary JSON in the mimebundles (for example, from ipywidgets), and we never flagged them before, but with nbformat package 5, they are now suddenly they are flagged as noncompliant.

This is a follow-up to https://github.com/jupyter/nbformat/pull/167, and extends the compatibility changes from 4.1 back to 4.0.