Fix installing package_data when --build-base is set (#211)
Fix package_data declarations in setup.py to associate every set
of files with its bottommost package.  This is necessary for the 'build'
command to install files correctly when --build-base is passed
explicitly.  The all-files-for-top-package approach seems to work only
incidentally.

To reproduce the problem, try:

    setup.py build --build-base=build2

and note that none of the data files were installed.