Fix and test auto version numbering.
Pep 440 mandate that the makers are `a`, `b`, `rc`, without dots.
Not beta/alpha. And `.dev` if present, must have a dot.

Otherwise this leads to problem in package resolutions where tgz can
appear to be more recent than whl of the same version...

This try to fix the version numbering scheme and test it.

It also skip whitespace trailing test on .bumpversoin.cfg at bum2version
itself  adds trailing spaces and try to commit which fails.